Speed up from reducing the number of castings

+++ AVC_transform.py
@@ -12,7 +12,7 @@ class AVC_transform:
         self.Cf = np.array([[1, 1, 1, 1],
                             [2, 1, -1, -2],
                             [1, -1, -1, 1],
-                            [1, -2, 2, -1]])
+                            [1, -2, 2, -1]], dtype=int)
         
         self.Ci = np.array([[1, 1, 1, 1],
                             [1, 1/2, -1/2, -1],
@@ -22,26 +22,26 @@ class AVC_transform:
         self.sec_Cf = np.array([[1, 1, 1, 1],
                                 [1, 1, -1, -1],
                                 [1, -1, -1, 1],
-                                [1, -1, 1, -1]])
+                                [1, -1, 1, -1]], dtype=int)
         
         self.sec_Ci = np.array([[1, 1, 1, 1],
                                 [1, 1, -1, -1],
                                 [1, -1, -1, 1],
-                                [1, -1, 1, -1]])
+                                [1, -1, 1, -1]], dtype=int)
         
         self.SM = np.array([[10, 16, 13],
                             [11, 18, 14],
                             [13, 20, 16],
                             [14, 23, 18],
                             [16, 25, 20],
-                            [18, 29, 23]])
+                            [18, 29, 23]], dtype=int)
         
         self.MF = np.array([[13107, 5243, 8066],
                             [11916, 4660, 7490],
                             [10082, 4194, 6554],
                             [9362, 3647, 5825],
                             [8192, 3355, 5243],
-                            [7282, 2893, 4559]])
+                            [7282, 2893, 4559]], dtype=int)
         
         if flag_uniform:
             self.base_Q = np.ones((4, 4))
@@ -89,7 +89,7 @@ class AVC_transform:
         V = np.array([[d, m, d, m],
                     [m, l, m, l],
                     [d, m, d, m],
-                    [m, l, m, l]]).astype(int)
+                    [m, l, m, l]])
 
         # Find the inverse quantized coefficients
         Wi = Z * V
@@ -98,7 +98,7 @@ class AVC_transform:
 
     def quantization(self, W, QP, ind=None):
         # q is qbits
-        q = 15 + np.floor(QP / 6).astype(int)
+        q = 15 + int(np.floor(QP / 6))
 
         # M is the multiplying factor which is found from QP value
         # MF is the multiplying factor matrix
@@ -122,10 +122,10 @@ class AVC_transform:
         M = np.array([[a, g, a, g],
                     [g, b, g, b],
                     [a, g, a, g],
-                    [g, b, g, b]])
+                    [g, b, g, b]], dtype=int)
 
         # Scaling and quantization
-        term = (W * M + 2**(q-1)).astype(int)
+        term = (W * M + 2**(q-1)).astype(int)  
         Z = (term >> q)
         return Z
